reorganize decorators and add comments
